Bump Openzeppelin version for Interchain wrapper

+++ src/infinity/interfaces/IWMATIC.sol
@@ -2,7 +2,7 @@
 
 pragma solidity ^0.8.16;
 
-import "oz/token/ERC20/IERC20.sol";
+import "oz-4-8-3/token/ERC20/IERC20.sol";
 
 /**
  * @author Cujo
@@ -12,5 +12,5 @@ import "oz/token/ERC20/IERC20.sol";
 interface IWMATIC is IERC20 {
     function deposit() external payable;
 
-    function withdraw(uint) external;
+    function withdraw(uint256) external;
 }
